+++ src/OG.Graphics.Abstraction/IOgGraphicsContext.cs
@@ -2,7 +2,7 @@
 namespace OG.Graphics.Abstraction;
 public interface IOgGraphicsContext
 {
-    Rect ClipRect   { get; set; }
-    Rect RenderRect { get; set; }
-    int  ZOrder     { get; set; }
+    Rect  ClipRect   { get; set; }
+    Rect  RenderRect { get; set; }
+    float ZOrder     { get; set; }
 }
