+++ src/OG.Graphics.Contexts/OgBaseGraphicsContext.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 namespace OG.Graphics.Contexts;
 public class OgBaseGraphicsContext : IOgGraphicsContext
 {
-    public Rect ClipRect   { get; set; }
-    public Rect RenderRect { get; set; }
-    public int  ZOrder     { get; set; }
+    public Rect  ClipRect   { get; set; }
+    public Rect  RenderRect { get; set; }
+    public float ZOrder     { get; set; }
 }
